現在の気温を表示するよう修正

+++ core/src/main/java/com/tatsuki/core/api/response/CurrentResponse.kt
@@ -21,7 +21,7 @@ data class CurrentResponse(
     val humidity: Int,
 
     @Json(name = "wind_speed")
-    val windSpeed: Int,
+    val windSpeed: Double,
 
     @Json(name = "weather")
     val weather: List<WeatherResponse>
