+++ core/src/main/java/com/tatsuki/core/entity/CurrentWeatherInfoItem.kt
@@ -3,5 +3,5 @@ package com.tatsuki.core.entity
 sealed class CurrentWeatherInfoItem {
     data class PressureItem(val value: Int) : CurrentWeatherInfoItem()
     data class HumidityItem(val value: Int) : CurrentWeatherInfoItem()
-    data class WindSpeedItem(val value: Int) : CurrentWeatherInfoItem()
+    data class WindSpeedItem(val value: Double) : CurrentWeatherInfoItem()
 }
